requirements.txt
# RAG-Enabled Chatbot - Requirements

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# Hugging Face & ML
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
sentence-transformers>=2.2.0

# RAG & Vector Search
faiss-cpu>=1.7.4
numpy>=1.24.0

# Document Processing
PyPDF2>=3.0.0
python-markdown>=3.5.0

# Data Validation
pydantic>=2.0.0

# Utilities
nltk>=3.8.0
regex>=2023.0.0

# Development Tools (optional)
# pytest>=7.4.0
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.6.0
